Move path canonicalization from API to kernel
