+++ src/api/process.rs
@@ -1,12 +1,9 @@
 use crate::api::syscall;
-use crate::api::fs;
 
 pub fn spawn(path: &str) -> Result<(), ()> {
-    if let Ok(path) = fs::canonicalize(path) {
-        if syscall::stat(&path).is_some() {
-            syscall::spawn(&path);
-            return Ok(());
-        }
+    if syscall::stat(&path).is_some() {
+        syscall::spawn(&path);
+        return Ok(());
     }
     Err(())
 }
